requirements.txt
astroid==2.2.5
autopep8==1.4.4
click==7.0
dnspython==1.16.0
flask-cors==3.0.7
flask-pymongo==2.2.0
flask==1.0.2
isort==4.3.17
itsdangerous==1.1.0
jinja2==2.10.1
lazy-object-proxy==1.3.1
markupsafe==1.1.1
mccabe==0.6.1
mongoengine==0.17.0
pycodestyle==2.5.0
pylint==2.3.1
pymongo==3.7.2
python-dateutil==2.8.0
six==1.12.0
tabulate==0.8.3
typed-ast==1.3.4 ; implementation_name == 'cpython'
werkzeug==0.15.2
wrapt==1.11.1
